// src/app/modules/openapi/model/check-out-order-dto.ts
/**
 * E-Commerce Application
 *
 * Contact: sad@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CheckOutOrderDTO { 
    address: string;
    payType: CheckOutOrderDTO.PayTypeEnum;
}
export namespace CheckOutOrderDTO {
    export const PayTypeEnum = {
        Cash: 'CASH',
        Credit: 'CREDIT'
    } as const;
    export type PayTypeEnum = typeof PayTypeEnum[keyof typeof PayTypeEnum];
}