// src/app/modules/openapi/model/user-dto.ts
/**
 * E-Commerce Application
 *
 * Contact: sad@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderDTO } from './order-dto';
import { AddressDTO } from './address-dto';
import { CartDTO } from './cart-dto';


export interface UserDTO { 
    userId?: number;
    userName?: string;
    email?: string;
    job?: string;
    creditNo?: string;
    creditLimit?: number;
    phone?: string;
    role?: UserDTO.RoleEnum;
    addresses?: Array<AddressDTO>;
    orders?: Array<OrderDTO>;
    carts?: Array<CartDTO>;
    bd?: string;
}
export namespace UserDTO {
    export const RoleEnum = {
        Admin: 'ADMIN',
        User: 'USER'
    } as const;
    export type RoleEnum = typeof RoleEnum[keyof typeof RoleEnum];
}